Rename sonic-dhcp-server
