+++ rules/docker-dhcp-server.mk
@@ -14,7 +14,7 @@ $(DOCKER_DHCP_SERVER)_DBG_IMAGE_PACKAGES = $($(DOCKER_CONFIG_ENGINE_BULLSEYE)_DB
 $(DOCKER_DHCP_SERVER)_LOAD_DOCKERS = $(DOCKER_CONFIG_ENGINE_BULLSEYE)
 
 $(DOCKER_DHCP_SERVER)_INSTALL_DEBS = $(PYTHON3_SWSSCOMMON)
-$(DOCKER_DHCP_SERVER)_PYTHON_WHEELS += $(SONIC_DHCPSERVD_PY3)
+$(DOCKER_DHCP_SERVER)_PYTHON_WHEELS += $(SONIC_DHCP_SERVER_PY3)
 
 SONIC_DOCKER_IMAGES += $(DOCKER_DHCP_SERVER)
 SONIC_DOCKER_DBG_IMAGES += $(DOCKER_DHCP_SERVER_DBG)
