+++ rules/sonic-dhcp-server.dep
@@ -1,5 +1,5 @@
 SPATH       := $($(SONIC_DHCPSERVD)_SRC_PATH)
-DEP_FILES   := $(SONIC_COMMON_FILES_LIST) rules/sonic-dhcpservd.mk rules/sonic-dhcpservd.dep
+DEP_FILES   := $(SONIC_COMMON_FILES_LIST) rules/sonic-dhcp-server.mk rules/sonic-dhcp-server.dep
 DEP_FILES   += $(SONIC_COMMON_BASE_FILES_LIST)
 SMDEP_FILES := $(addprefix $(SPATH)/,$(shell cd $(SPATH) && git ls-files))
