+++ rules/sonic-dhcp-server.mk
@@ -1,0 +1,10 @@
+# sonic-dhcp-server package
+
+SONIC_DHCP_SERVER_PY3 = sonic_dhcp_server-1.0-py3-none-any.whl
+$(SONIC_DHCP_SERVER_PY3)_SRC_PATH = $(SRC_PATH)/sonic-dhcp-server
+$(SONIC_DHCP_SERVER_PY3)_DEPENDS += $(SONIC_PY_COMMON_PY3)
+$(SONIC_DHCP_SERVER_PY3)_DEBS_DEPENDS = $(LIBSWSSCOMMON) $(PYTHON3_SWSSCOMMON)
+$(SONIC_DHCP_SERVER_PY3)_PYTHON_VERSION = 3
+ifeq ($(INCLUDE_DHCP_SERVER), y)
+SONIC_PYTHON_WHEELS += $(SONIC_DHCP_SERVER_PY3)
+endif
