--- rules/sonic-dhcpservd.mk
@@ -1,10 +1,0 @@
-# sonic-dhcpservd package
-
-SONIC_DHCPSERVD_PY3 = sonic_dhcpservd-1.0-py3-none-any.whl
-$(SONIC_DHCPSERVD_PY3)_SRC_PATH = $(SRC_PATH)/sonic-dhcpservd
-$(SONIC_DHCPSERVD_PY3)_DEPENDS += $(SONIC_PY_COMMON_PY3)
-$(SONIC_DHCPSERVD_PY3)_DEBS_DEPENDS = $(LIBSWSSCOMMON) $(PYTHON3_SWSSCOMMON)
-$(SONIC_DHCPSERVD_PY3)_PYTHON_VERSION = 3
-ifeq ($(INCLUDE_DHCP_SERVER), y)
-SONIC_PYTHON_WHEELS += $(SONIC_DHCPSERVD_PY3)
-endif
