+++ src/sonic-dhcp-server/setup.cfg
@@ -7,7 +7,7 @@ testpaths = tests
 
 [coverage:run]
 branch = True
-source = dhcpservd
+source = dhcp_server
 
 [coverage:report]
 exclude_lines =
