+++ src/sonic-dhcp-server/tests/test_data/mock_config_db_without_port_config.json
